+++ internal/work/givescan.go
@@ -28,7 +28,7 @@ var scanSecondPart float64
 func GiveScanInfo(p packet.Packet, conn net.Conn) (task.TaskResult, error) {
 	key := p.GetRkey()
 	logger.Info("GiveScanInfo: ", zap.Any("message", key+", Msg: "+p.GetMessage()))
-	scanFirstPart = config.Viper.GetFloat64("SCAN_FIRST_PART")
+	scanFirstPart = float64(config.Viper.GetInt("SCAN_FIRST_PART"))
 	scanSecondPart = 100 - scanFirstPart
 	total, err := strconv.Atoi(p.GetMessage())
 	if err != nil {
